Corretti alcuni errori su compressore e decompressore.

+++ decompressor.h
@@ -13,7 +13,7 @@
 struct elem{
 	
 	char c;  
-	int parent; 
+	uint64_t parent;
 	//child id è la sua posizione nell'array
 };
